Move correlative workflow related endpoints
