+++ src/murfey/util/models.py
@@ -73,40 +73,6 @@ class RsyncerInfo(BaseModel):
 
 
 """
-FIB
-===
-Models related to FIB, as part of correlative workflow with TEM.
-"""
-
-
-class Sample(BaseModel):
-    sample_group_id: int
-    sample_id: int
-    subsample_id: int
-    image_path: Optional[Path]
-
-
-class BLSampleImageParameters(BaseModel):
-    sample_id: int
-    sample_path: Path
-
-
-class BLSampleParameters(BaseModel):
-    sample_group_id: int
-
-
-class BLSubSampleParameters(BaseModel):
-    sample_id: int
-    image_path: Optional[Path] = None
-
-
-class MillingParameters(BaseModel):
-    lamella_number: int
-    images: List[str]
-    raw_directory: str
-
-
-"""
 Single Particle Analysis
 ========================
 Models related to the single-particle analysis workflow.
